[google_maps_flutter] Fix scaling for mipmap markers
In the mipmap case, the Dart side of the Maps plugin was passing in the asset image scale as the scale factor, discarding the actual device pixel ratio and producing a scale factor with an inverse effect from the non-mipmapped case.

For web, set the device pixel ratio to 1 since JS Maps will display markers in logical pixels.